requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
asgi-lifespan==2.1.0
email-validator==2.1.1
fastapi==0.111.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
pytest==8.2.0
pytest-asyncio==0.23.6
python-jose==3.3.0
python-multipart==0.0.9
sqlalchemy==2.0.29
uvicorn==0.30.1
